codepoint: change return type of cp_len() to int
Since cp_len() will never return anything greater than 4 (maybe 6 at some point)
using size_t only provokes warnings and forces additional includes.
int8_t isn't used because that might cause trouble with some functions
interpreting it as an ASCII value and not a number.

+++ include/encode.h
@@ -4,12 +4,11 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#include <sys/types.h>
 
 typedef uint_least32_t Codepoint;
 
 /* codepoint.c */
 bool cp_isvalid(Codepoint);
-size_t cp_len(Codepoint);
+int cp_len(Codepoint);
 
 #endif /* ENCODE_H_ */
